samples: get rid of warnings from common files.

+++ samples/common/common.cpp
@@ -150,18 +150,18 @@ void SlotFinder::aggregate() {
     uint32_t offsetToOR_last = m_offsetsToOR_AND[2 * 0 + 0];
     uint32_t offsetToAND_last = m_offsetsToOR_AND[2 * 0 + 1];
     uint32_t offsetToNumUsedFlags_last = m_offsetsToNumUsedFlags[0];
-    for (int layer = 1; layer < m_numLayers; ++layer) {
+    for (int layer = 1; layer < static_cast<int32_t>(m_numLayers); ++layer) {
         uint32_t numFlagBinsInLayer = nextMultiplierForPowOf2(m_numFlagsInLayerList[layer], 5);
         uint32_t offsetToOR = m_offsetsToOR_AND[2 * layer + 0];
         uint32_t offsetToAND = m_offsetsToOR_AND[2 * layer + 1];
         uint32_t offsetToNumUsedFlags = m_offsetsToNumUsedFlags[layer];
-        for (int binIdx = 0; binIdx < numFlagBinsInLayer; ++binIdx) {
+        for (int binIdx = 0; binIdx < static_cast<int32_t>(numFlagBinsInLayer); ++binIdx) {
             uint32_t &ORFlagBin = m_flagBins[offsetToOR + binIdx];
             uint32_t &ANDFlagBin = m_flagBins[offsetToAND + binIdx];
             uint32_t &numUsedFlagsUnderBin = m_numUsedFlagsUnderBinList[offsetToNumUsedFlags + binIdx];
 
             uint32_t numFlagsInBin = std::min(32u, m_numFlagsInLayerList[layer] - 32 * binIdx);
-            for (int bit = 0; bit < numFlagsInBin; ++bit) {
+            for (int bit = 0; bit < static_cast<int32_t>(numFlagsInBin); ++bit) {
                 uint32_t lBinIdx = 32 * binIdx + bit;
                 uint32_t lORFlagBin = m_flagBins[offsetToOR_last + lBinIdx];
                 uint32_t lANDFlagBin = m_flagBins[offsetToAND_last + lBinIdx];
@@ -188,7 +188,7 @@ void SlotFinder::resize(uint32_t numSlots) {
     newFinder.initialize(numSlots);
 
     uint32_t numLowestFlagBins = std::min(m_numLowestFlagBins, newFinder.m_numLowestFlagBins);
-    for (int binIdx = 0; binIdx < numLowestFlagBins; ++binIdx) {
+    for (int binIdx = 0; binIdx < static_cast<int32_t>(numLowestFlagBins); ++binIdx) {
         uint32_t numFlagsInBin = std::min(32u, numSlots - 32 * binIdx);
         uint32_t mask = numFlagsInBin >= 32 ? 0xFFFFFFFF : ((1 << numFlagsInBin) - 1);
         uint32_t value = m_flagBins[0 + binIdx] & mask;
@@ -207,7 +207,7 @@ void SlotFinder::setInUse(uint32_t slotIdx) {
 
     bool setANDFlag = false;
     uint32_t flagIdxInLayer = slotIdx;
-    for (int layer = 0; layer < m_numLayers; ++layer) {
+    for (int layer = 0; layer < static_cast<int32_t>(m_numLayers); ++layer) {
         uint32_t binIdx = flagIdxInLayer / 32;
         uint32_t flagIdxInBin = flagIdxInLayer % 32;
 
@@ -234,7 +234,7 @@ void SlotFinder::setNotInUse(uint32_t slotIdx) {
 
     bool resetORFlag = false;
     uint32_t flagIdxInLayer = slotIdx;
-    for (int layer = 0; layer < m_numLayers; ++layer) {
+    for (int layer = 0; layer < static_cast<int32_t>(m_numLayers); ++layer) {
         uint32_t binIdx = flagIdxInLayer / 32;
         uint32_t flagIdxInBin = flagIdxInLayer % 32;
 
@@ -308,7 +308,7 @@ uint32_t SlotFinder::find_nthUsedSlot(uint32_t n) const {
     for (int layer = m_numLayers - 1; layer >= 0; --layer) {
         uint32_t numUsedFlagsOffset = m_offsetsToNumUsedFlags[layer];
         uint32_t numFlagBinsInLayer = nextMultiplierForPowOf2(m_numFlagsInLayerList[layer], 5);
-        for (int binIdx = startBinIdx; binIdx < numFlagBinsInLayer; ++binIdx) {
+        for (int binIdx = startBinIdx; binIdx < static_cast<int32_t>(numFlagBinsInLayer); ++binIdx) {
             uint32_t numUsedFlagsUnderBin = m_numUsedFlagsUnderBinList[numUsedFlagsOffset + binIdx];
 
             // JP: 現在のビンの配下にインデックスnの使用中スロットがある。
@@ -332,7 +332,7 @@ uint32_t SlotFinder::find_nthUsedSlot(uint32_t n) const {
 void SlotFinder::debugPrint() const {
     uint32_t numLowestFlagBins = nextMultiplierForPowOf2(m_numFlagsInLayerList[0], 5);
     hpprintf("----");
-    for (int binIdx = 0; binIdx < numLowestFlagBins; ++binIdx) {
+    for (int binIdx = 0; binIdx < static_cast<int32_t>(numLowestFlagBins); ++binIdx) {
         hpprintf("------------------------------------");
     }
     hpprintf("\n");
@@ -340,13 +340,13 @@ void SlotFinder::debugPrint() const {
         hpprintf("layer %u (%u):\n", layer, m_numFlagsInLayerList[layer]);
         uint32_t numFlagBinsInLayer = nextMultiplierForPowOf2(m_numFlagsInLayerList[layer], 5);
         hpprintf(" OR:");
-        for (int binIdx = 0; binIdx < numFlagBinsInLayer; ++binIdx) {
+        for (int binIdx = 0; binIdx < static_cast<int32_t>(numFlagBinsInLayer); ++binIdx) {
             uint32_t ORFlagBin = m_flagBins[m_offsetsToOR_AND[2 * layer + 0] + binIdx];
             for (int i = 0; i < 32; ++i) {
                 if (i % 8 == 0)
                     hpprintf(" ");
 
-                bool valid = binIdx * 32 + i < m_numFlagsInLayerList[layer];
+                bool valid = binIdx * 32 + i < static_cast<int32_t>(m_numFlagsInLayerList[layer]);
                 if (!valid)
                     continue;
 
@@ -356,13 +356,13 @@ void SlotFinder::debugPrint() const {
         }
         hpprintf("\n");
         hpprintf("AND:");
-        for (int binIdx = 0; binIdx < numFlagBinsInLayer; ++binIdx) {
+        for (int binIdx = 0; binIdx < static_cast<int32_t>(numFlagBinsInLayer); ++binIdx) {
             uint32_t ANDFlagBin = m_flagBins[m_offsetsToOR_AND[2 * layer + 1] + binIdx];
             for (int i = 0; i < 32; ++i) {
                 if (i % 8 == 0)
                     hpprintf(" ");
 
-                bool valid = binIdx * 32 + i < m_numFlagsInLayerList[layer];
+                bool valid = binIdx * 32 + i < static_cast<int32_t>(m_numFlagsInLayerList[layer]);
                 if (!valid)
                     continue;
 
@@ -372,7 +372,7 @@ void SlotFinder::debugPrint() const {
         }
         hpprintf("\n");
         hpprintf("    ");
-        for (int binIdx = 0; binIdx < numFlagBinsInLayer; ++binIdx) {
+        for (int binIdx = 0; binIdx < static_cast<int32_t>(numFlagBinsInLayer); ++binIdx) {
             uint32_t numUsedFlagsUnderBin = m_numUsedFlagsUnderBinList[m_offsetsToNumUsedFlags[layer] + binIdx];
             hpprintf("                            %8u", numUsedFlagsUnderBin);
         }
@@ -382,13 +382,13 @@ void SlotFinder::debugPrint() const {
         hpprintf("layer 0 (%u):\n", m_numFlagsInLayerList[0]);
         uint32_t numFlagBinsInLayer = nextMultiplierForPowOf2(m_numFlagsInLayerList[0], 5);
         hpprintf("   :");
-        for (int binIdx = 0; binIdx < numFlagBinsInLayer; ++binIdx) {
+        for (int binIdx = 0; binIdx < static_cast<int32_t>(numFlagBinsInLayer); ++binIdx) {
             uint32_t ORFlagBin = m_flagBins[binIdx];
             for (int i = 0; i < 32; ++i) {
                 if (i % 8 == 0)
                     hpprintf(" ");
 
-                bool valid = binIdx * 32 + i < m_numFlagsInLayerList[0];
+                bool valid = binIdx * 32 + i < static_cast<int32_t>(m_numFlagsInLayerList[0]);
                 if (!valid)
                     continue;
 
@@ -398,7 +398,7 @@ void SlotFinder::debugPrint() const {
         }
         hpprintf("\n");
         hpprintf("    ");
-        for (int binIdx = 0; binIdx < numFlagBinsInLayer; ++binIdx) {
+        for (int binIdx = 0; binIdx < static_cast<int32_t>(numFlagBinsInLayer); ++binIdx) {
             uint32_t numUsedFlagsUnderBin = m_numUsedFlagsUnderBinList[binIdx];
             hpprintf("                            %8u", numUsedFlagsUnderBin);
         }
@@ -421,8 +421,8 @@ void saveImage(const std::filesystem::path &filepath, uint32_t width, uint32_t h
 void saveImage(const std::filesystem::path &filepath, uint32_t width, uint32_t height, const float4* data,
                bool applyToneMap, bool apply_sRGB_gammaCorrection) {
     auto image = new uint32_t[width * height];
-    for (int y = 0; y < height; ++y) {
-        for (int x = 0; x < width; ++x) {
+    for (int y = 0; y < static_cast<int32_t>(height); ++y) {
+        for (int x = 0; x < static_cast<int32_t>(width); ++x) {
             float4 src = data[y * width + x];
             if (applyToneMap) {
                 src.x = simpleToneMap_s(src.x);
@@ -435,10 +435,10 @@ void saveImage(const std::filesystem::path &filepath, uint32_t width, uint32_t h
                 src.z = sRGB_gamma_s(src.z);
             }
             uint32_t &dst = image[y * width + x];
-            dst = ((std::min<uint32_t>(src.x * 255, 255) << 0) |
-                   (std::min<uint32_t>(src.y * 255, 255) << 8) |
-                   (std::min<uint32_t>(src.z * 255, 255) << 16) |
-                   (std::min<uint32_t>(src.w * 255, 255) << 24));
+            dst = ((std::min<uint32_t>(static_cast<uint32_t>(src.x * 255), 255) << 0) |
+                   (std::min<uint32_t>(static_cast<uint32_t>(src.y * 255), 255) << 8) |
+                   (std::min<uint32_t>(static_cast<uint32_t>(src.z * 255), 255) << 16) |
+                   (std::min<uint32_t>(static_cast<uint32_t>(src.w * 255), 255) << 24));
         }
     }
